+++ empire/server/plugins/websockify_server/websockify_server.py
@@ -13,6 +13,7 @@ log = logging.getLogger(__name__)
 class Plugin(BasePlugin):
     @override
     def on_load(self, db):
+        self.execution_enabled = False
         self.csharpserver_proc = None
         self.settings_options = {
             "SourceHost": {